# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenEnv AI Web Navigation Training Environment API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ResetRequest {
  /** Optional task ID to use for this session (defaults to random) */
  taskId?: string;
  /** Optional random seed for reproducibility */
  seed?: number;
}

/**
 * Action parameters (e.g., product_name for search_product)
 */
export type StepRequestParams = { [key: string]: unknown };

export interface StepRequest {
  /** Action to execute (open_home, search_product, click_product, add_to_cart, noop) */
  action: string;
  /** Action parameters (e.g., product_name for search_product) */
  params?: StepRequestParams;
}

/**
 * Additional info about the step
 */
export type StepResultInfo = { [key: string]: unknown };

export interface Product {
  id: string;
  name: string;
  price: number;
  category: string;
  description?: string;
  rating?: number;
}

export interface Observation {
  /** Current page the agent is on (home, search_results, product_detail, cart) */
  page: string;
  /** List of actions available from current state */
  available_actions: string[];
  /** Products visible in search results */
  search_results?: Product[];
  /** Currently viewed product (if on product page) */
  current_product?: Product;
  /** Items currently in cart */
  cart_items?: Product[];
  /** Description of the current task */
  task_description: string;
  /** ID of the current task */
  task_id: string;
  /** Number of steps taken in this episode */
  step_count: number;
  /** Cumulative reward so far */
  total_reward: number;
}

export interface StepResult {
  observation: Observation;
  /** Reward received for this step (normalized 0.0-1.0) */
  reward: number;
  /** Whether the episode is complete */
  done: boolean;
  /** Additional info about the step */
  info?: StepResultInfo;
}

export interface EnvironmentState {
  observation: Observation;
  session_id: string;
  is_active: boolean;
  created_at?: string;
}

export type TaskDifficulty =
  (typeof TaskDifficulty)[keyof typeof TaskDifficulty];

export const TaskDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type TaskRewardBreakdown = { [key: string]: unknown };

export interface Task {
  id: string;
  name: string;
  description: string;
  difficulty: TaskDifficulty;
  target_product: string;
  max_steps: number;
  reward_breakdown?: TaskRewardBreakdown;
}

export interface TaskList {
  tasks: Task[];
}

export interface SessionRecord {
  session_id: string;
  task_id: string;
  task_name: string;
  total_reward: number;
  steps_taken: number;
  completed: boolean;
  grader_score: number;
  started_at: string;
  ended_at?: string;
}

export interface SessionHistory {
  sessions: SessionRecord[];
  total_sessions: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}