# lib/api-zod/src/generated/types/environmentState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenEnv AI Web Navigation Training Environment API
 * OpenAPI spec version: 0.1.0
 */
import type { Observation } from "./observation";

export interface EnvironmentState {
  observation: Observation;
  session_id: string;
  is_active: boolean;
  created_at?: string;
}