# lib/api-zod/src/generated/types/observation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenEnv AI Web Navigation Training Environment API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface Observation {
  /** Current page the agent is on (home, search_results, product_detail, cart) */
  page: string;
  /** List of actions available from current state */
  available_actions: string[];
  /** Products visible in search results */
  search_results?: Product[];
  /** Currently viewed product (if on product page) */
  current_product?: Product;
  /** Items currently in cart */
  cart_items?: Product[];
  /** Description of the current task */
  task_description: string;
  /** ID of the current task */
  task_id: string;
  /** Number of steps taken in this episode */
  step_count: number;
  /** Cumulative reward so far */
  total_reward: number;
}