# lib/api-zod/src/generated/types/resetRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenEnv AI Web Navigation Training Environment API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetRequest {
  /** Optional task ID to use for this session (defaults to random) */
  taskId?: string;
  /** Optional random seed for reproducibility */
  seed?: number;
}