# lib/api-zod/src/generated/types/sessionHistory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenEnv AI Web Navigation Training Environment API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionRecord } from "./sessionRecord";

export interface SessionHistory {
  sessions: SessionRecord[];
  total_sessions: number;
}