# lib/api-zod/src/generated/types/sessionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenEnv AI Web Navigation Training Environment API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionRecord {
  session_id: string;
  task_id: string;
  task_name: string;
  total_reward: number;
  steps_taken: number;
  completed: boolean;
  grader_score: number;
  started_at: string;
  ended_at?: string;
}