# lib/api-zod/src/generated/types/stepRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenEnv AI Web Navigation Training Environment API
 * OpenAPI spec version: 0.1.0
 */
import type { StepRequestParams } from "./stepRequestParams";

export interface StepRequest {
  /** Action to execute (open_home, search_product, click_product, add_to_cart, noop) */
  action: string;
  /** Action parameters (e.g., product_name for search_product) */
  params?: StepRequestParams;
}