# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenEnv AI Web Navigation Training Environment API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskDifficulty } from "./taskDifficulty";
import type { TaskRewardBreakdown } from "./taskRewardBreakdown";

export interface Task {
  id: string;
  name: string;
  description: string;
  difficulty: TaskDifficulty;
  target_product: string;
  max_steps: number;
  reward_breakdown?: TaskRewardBreakdown;
}